Resolves #1639: Support a configurable FDB API version
This makes the FDB API version configurable through the `FDBDatabaseFactory`.

We'll need some better way of testing this, I think. I've tried running the tests at both the 6.3 and 7.1 API versions, and it passed, but this value cannot be changed after the client has been initiated, which makes this somewhat complex to test with our normal approach. We might be able to use a test-only implementation of `FDBDatabaseFactory` if we wanted to, say, test future work that had feature flags on the API version. Another approach might be to run the tests multiple times, once per API version

This resolves #1639.