Fixes #1643: Remove JCenter from build repositories
This removes the references to `jcenter()` from our gradle configuration. There was one dependency, one of the plugins, that didn't seem to have an up-to-date version on `mavenCentral()` (which was potentially why we were using `jcenter()` for anything), but they are now in the `gradlePluginsPortal()`, and so I switched those over to using the new gradle plugins format. The two remaining plugins did not seem to have entries in the `gradlePluginsPortal()`, so I left them as is for now.

I was able to redownload the dependencies locally, but a CI run would be good to know to double check that there wasn't accidentally some cache that made the build possible.