Resolves #1636: Increase FDB Java version to FDB 7.1  (#1641)
* Resolves #1636: Increase FDB Java version to FDB 7.1

This increases the FDB Java dependency to FDB 7.1. This keeps the API version at 630 so that the library can still be used to connect to FDB 6.3 and FDB 7.0 clusters, though it would now require using a 7.1 FDB C client as the primary client with the other clients as external clients.

This will allow for additional features to be built on top of FDB 7.1, though we probably want to also complete #1639 so that we can responsibly guard features reliant on 7.1 in environments where the cluster may still be running older FDB server versions.

* Resolves #1640: Dependency bump for 3.2 release

This updates the proto3 version to 3.20.1 and updates a few other dependencies as well. It leaves the guava version at 30.1-jre instead of updating it to 31.1-jre because that causes spotbugs violations regarding nullability problems. I think those errors are because of a possible change in the way Guava handles nullability annotations, but it seems like something we should fix later.

* fix capitalization in documentation