Fixes #1657: build error on Apple silicon with PB 2.6 (#1658)
As suggested in the issue, on macOS Apple Silicon architectures, this reverts to an x86_64 version of the proto2 compiler for compatibility reasons. This will run in emulation on machines with that architecture. Note also that we do not ship protoc with our binaries, nor do we link against it, so there is not a concern that this would cause problems at link time for software compiled for Apple Silicon. The protoc binary is only used during our own compilation to generate Java files from our `.proto` files.

This fixes #1657.