Make suggested TransactionalLimitedRunner constructor take FDBRecordContextConfig
I don't think I can just refuse to take a builder, because indexing
passes in the builder from the runner, so it's possible that there
is a reason to change the runner while an index is being built.
There's no thread safety in that builder though, so I doubt it would
work perfectly if you actually tried to do that.