Fix FDBRecordStoreIndexTest using MockedLocalityUtil
I don't know how this is related, but it clearly started failing
with the new runner because it was using the real locality,
instead of the mocked one. I fixed it to ensure that it uses
the mocked one.
Perhaps it's because the ordering of tests being run changed and
with the other changes it now runs some other test that accesses
the database before this test, which causes the test to use a
cached database.