Resolves #1692: Remove ERROR level log on failed no-ops
This removes the `ERROR`-level log on no-op failure. The error was already getting propagated, and that, in practice, should be enough for adopters to detect errors and (if they so desire) log the error. I considered adding logic to include the `cluster` as logging details, but that ended up being a little bit delicate because it only works if the underlying failure is a `LoggableException`, which isn't necessarily the case because we haven't called the error-wrapping logic at that point. So, it was easier to just leave it without that information, and it didn't seem like a huge loss, as the caller should be able to determine that based on the fact that they need an `FDBDatabase` to call this on anyway.

This resolves #1692.