Resolves #1743: Update FDB dependency to 7.1.10
The FDB Java version has been updated to 7.1.10 in this PR. The more important thing is actually that the client (and server) versions be updated to 7.1.10 or 7.1.11 (depending on whether the environments support AVX instructions) to get the recent bug fixes, but this seems like a good signal to send to let our downstream adopters know they should upgrade.

This resolves #1743.