Resolves #1714: LocatableResolver::reverseLookup should support borrowing read versions
This adds arguments to `reverseLookup` to allow a transaction to be included that can be used to get a read version, much like the `resolve` methods. This means that operations that resolve multiple entries from the reverse directory cache can avoid needing to get a read version for each transaction, which can add significant latency to such requests.

This resolves #1714.