Resolves #1758: Add index scan plan with additional reads to load the RYW cache
This adds a new `RecordQueryPlan` that has the same semantics (in terms of results returned, including continuations) as a `RecordQueryIndexPlan`, but that scans additional ranges for the purposes
of loading the FDB C client's read-your-writes cache. If the query is executed multiple times in the same transaction with different bound parameters, then loading this cache might result in performance
wins because query executions (after the first) can use the cache to avoid having to make additional requests to the database. At the same time, because FDB paginates range scans, the amount of additional
work that each scan has to do to load the cache should be bounded as well (to approximately one page of results per scan).

This resolves #1758.