Support singleton tables (REL-190)
* Support singleton tables

* Add SINGLE ROW ONLY syntax and add more tests

Record Layer supports a singleton table whose primary key consists only of a RecordTypeKey.
In Relational parlance, that translates to a table that does not "seem" to declare a primary key at all.

This PR introduces a new syntax to define singleton tables:

CREATE TABLE A(a int64, SINGLE ROW ONLY)
Tables now need to either specify SINGLE ROW ONLY or explicitly define a PRIMARY KEY.

On Singleton tables, executeGet and executeDelete can now be called with an empty KeySet.