Resolves #1857: DirectoryLayerDirectory creates new transactions during reverse lookup
This fixes up the `DirectoryLayerDirectory` to make use of the `LocatableResolver::reverseLookup` variant that takes an `FDBRecordContext` instead of just a timer. The latter was added as part of #1714 so that lookups through the reverse directory layer could leverage an already initialized transaction's read version. The data here is write-once, so it is safe to read stale data (as long as the data aren't absent, and that case is retried with a new read version internally). This should decrease the amount of time it takes to list a key space path with values in the directory layer, as now, the operation doesn't need to create a new transaction if an initialized one already exists.

This resolves #1857.