Fixes #1876: AndWithThenPlanner does not handle OneOfThemWithComponent expressions
This adds support to the old planner to address a shortcoming where certain queries were choosing to use less efficient indexes due to a shortcoming with the `AndWithThenPlanner`. Based on the tests in the `NestedFieldQueryTest`, it looks like the new planner was already matching such expressions in this case, and so this shortcoming should only be in the old planner.

This fixes #1876.