(Enum Type support in relational) (REL-247)
This adds support for enum types in the DDL syntax, and it adds support for enums through the ResultSet and Row abstractions.

Within the DDL, an enum is created with syntax like:

```
CREATE ENUM suit ('SPADES', 'HEARTS', 'DIAMONDS', 'CLUBS');
```

Enums can then be referenced by name elsewhere in the DDL as types for fields in tables and structs.

This gets translated to protobuf by creating an enum with the given names as values. The order of the enum value in the DDL determines the protobuf value number which also means it will determine the sort order (as is in line with how Postgres sort order).

When accessing an enum via a result set or row, the user interacts with it as a string. In theory, we may want to one day consider exposing the ordinal variant as well, which would support renames, though at the expense of human readability.