Resolves #1942: Make unrolled deletes the default record delete behavior
This changes the default value in the property controlling whether record deletes are unrolled. Adopters can still opt into the old behavior by setting the property value, though most use cases should be fine with the new default. At some point in the future, we can consider deprecating the property and always unrolling record deletes if we wish.

This resolves #1942.