(JDBC Client: implement direct access api) (REL-312)
Implement Direct Access API in jdbc client; added insert, get, and scan.

Remove the Spanner ResultSet proto files and replace with a simpler ResultSet,
a derivative, that allows passing extra metadata for Struct and Array and that
enforces an ordering on Struct and Array metadata (facility not
available in Spanner ResultSet).

Make a testFixture in fdb-relational-core that exposes all of the assert
utility so can use it downstream in relational-jdbc to prove our returned
ResultSets are whole. Also add to testFixture means of obtaining
protobufs to use in DAAPI insert (so can avoid having to have client
jdbc have to know about DynamicMessageBuilder, et al.)

Add a version of SimpleDirectAccessInsertionTests on the client jdbc
side that runs the same test as the SimpleDirectAccessInsertionTests
that is in fdb-relational-core to exercise the new Direct Access API.

Remove UNKNOWNStatusInterceptor. It was not comprehensive enough. Played
more with tooling around interceptors and returning server-side error to
client; all is better now but a WIP still.

Signed-off-by: brines