Resolves #1973: Clearing index data can skip uniqueness violations space on non-unique indexes
When clearing out index data, we don't really need to clear out the list of uniqueness violations on non-unique indexes, as that data is only consulted by unique indexes. This skips clearing out that space for those indexes, and it also adjusts the `scan` method to only ask unique indexes for violations.

This resolves #1973.