Resolves #1967: Consider alternative mechanism of loading records which favors gets over scans
This adds a property-controlled method for reading single records from FDB. It issues multiple single-key gets instead of one scan (for reasons detailed in #1967). This is off by default, because it is expected that this will perform worse on B-tree based storage engines (including the default SSD storage engine), and it is highly experimental. But once in, users can adjust this configuration to configure how reads are orchestrated in their own experiments.

This resolves #1967.