decrease flakiness in `testVersionChangedFlagAfterOpening` which failed during prb
The error, as far as I can tell, ended up being that we have a static `Index` object, but those can be modified when adding to a meta-data object, which is what the test was doing. This fixes that, but it also cleans up some other things related to the test, namely that it was clearing out the meta-data space but not using it and was creating futures but never waiting on them (which both meant that they could continue past the transaction and that any asserts could be dropped), and there were some things like marking a `boolean` function and method with nullability annotations.