Resolves #1982: LOAD_RECORDS_VIA_GETS should default to false (#1983)
This sets the default value of the `LOAD_RECORD_VIA_GETS` property to false. This reverts the default behavior to match what the behavior prior to 3.3.327.0 when the property was introduced. It was always the intention that this property would be off by default, and then adopters could experiment with it to see if it improved performance, but the default was set incorrectly in the initial commit.

This resolves #1982.