Resolves #1995: Instrument RangeSet operations during index builds
This adds additional metrics for the RangeSet class. This refactors access of the class so that it goes through a wrapping object, which associates a RangeSet with a particular FDBRecordContext. That context is then used to supply the transaction for the range set, but it also is used to instrument operations of this new wrapper class so that we have timing information for RangeSet operations during an index build.

This resolves #1995.