Resolves #1998: Delay publishing commit-dependent metrics until after commit
This allows `FDBStoreTimer.Event`s to mark themselves as delayed until commit, in which case the context won't increment/record them initially, and they'll only get added to the base timer once the transaction has successfully committed. This is done by having the record context keep track of two timers, one of which is from the user and is updated immediately and the other of which is created by the context and stores these delayed metrics. Once the commit happens, the data from the delayed timer are added to the other timer.

I'd initially wanted this to be done with a single timer, which would then encapsulate the dispatching between the two timers. I think with a slightly different class hierarchy (such as a single `StoreTimer` interface with various implementations arranged to support composability), this could be done. However, because this uses inheritance instead of composition, it would have been a bit clunky and ultimately error prone.

This resolves #1998.