Fixes #2002: "EMPTY_SCANS" metric double counts empty ranges if hasNext() called multiple times
At the end of a range scan, it is legal to call `hasNext()` mutliple times. The old code would allow `hasNext()` to be called multiple times, and then it would increment the `EMPTY_SCANS` counter each time, which could result in the metric appearing the exceed the `RANGE_READS` metric. This also fixes another way in which `RANGE_READS` could be undercounted, as it is legal to start multiple `AsyncIterator`s from one `AsyncIterable`, though I don't think we actually do that within the Record Layer. Those two fixes together should ensure that `EMPTY_SCANS` is never higher than `RANGE_READS`.

This fixes #2002.