Fixes #2024: Listing missing ranges in a RangeSet has overly large conflict ranges
This adjusts the way that the `RangeSet` adds conflict ranges when listing missing ranges so that it only adds a read conflict range if the range is relevant to the final answer. This means that, for example, disjoint ranges can be modified in concurrent transacitons and all commit. There was already logic that adjusted these ranges in `insertRange`, so only the list method needed to be adjusted.

This fixes #2024.