Fixes #2015: ValueIndexExpansionVisitor does not handle when primary key subsumed by index key (#2016)
This adds a check while adding the primary key fields to the root expression so that if the primary key is entirely contained within the index expression and therefore doesn't need to be added to the key expression. There's a new test of the planner that fails without this fix, as the new planner needs to consider an expression of that kind. The test also runs under the old planner, which succeeded before and after this fix.

This fixes #2015.