(add an option to disable creation of record type discriminator as part of the primary key) (REL-369)
*  (add an option to disable creation of record type discriminator as part of the primary key)

This adds a `SchemaTemplateOption` that controls whether the record type key is added to the primary key of a record type when it is created. A pre-existing test around `DeleteRange` with no record-type key prefix was modified to now use the new option, and it validates that the record type key is omitted with the option set. I also added a new test class with some basic tests, though we probably want more coverage.

The customer yaml tests are also almost ready to use the new option. There were, however, a few problems that became clear when the option was set. One of them is that the Cascades planner in the Record Layer cannot handle value indexes where the full primary key is in the index entry, though that's an easy fix (that will need a Record Layer change). There was also one plan that changed, though we're workin on understanding that. But most of the queries worked as expected.

The new/modified tests do show that there's a weird interaction between some of the direct access APIs and this new option. In particular, for methods like `executeGet`, the table parameter might not be respected if the record at the given key is of the wrong type. This manifests in a bit of a weird way (as the type of the record on disk is stuffed into the declared table). Similar concerns are to be had with the `deleteRange` method. We could consider introducing an `ANY` table (or something like that) to the direct access methods to handle the fact that sometimes we don't know the table a row will belong to.