Resolves #2053: Default protoMajorVersion to 3 in build
This updates the build script so that if one runs `./gradlew build` with no other arguments, it will build the proto3-based artifacts. If one sets the `PROTO_VERSION` to 2, it should still build the proto2-based artifacts, so this won't change the published artifacts, but it may affect flows of a developer working on the project.

This resolves #2053.