(Plan cache always misses) (REL-385)
*  (Plan cache always misses)

This fixes the problem with the plan cache that was identified in the issue, which is that it was not identifying the set of readable indexes correctly. In particular, the `RecordStoreState` only contains a set of _non_-`READABLE` indexes, so to get a set of all `READABLE` indexes, it needs to combine the list of indexes from the meta-data with the record store state.

It also needs to look at a slightly different range of keys in the map in order to find a compatible plan. In particular, the cache contains elements keyed by their query and the sorted list of indexes that they rely on. Therefore, for a given query and a given store, to find all of the possible candidates, it needs to scan a range from `(query, [])` to `(query, [maxReadableIndex])` where `maxReadableIndex` is the `READABLE` index with the name that sorts last. The low endpoint is hopefully self-explanatory, but the high endpoint is less obvious. Effectively, it is safe to exclude all entries in the cache greater than `(query, [maxReadableIndex])` because either the entry will be for a different query, or it will contain a non-`READABLE` index in its "used index set". This follows because a list of indexes greater than `[maxReadableIndex]` will either begin with an element that sorts larger than `maxReadableIndex` (which must not be `READABLE` as `maxReadableIndex` is the largest `READABLE` index) or it will be a list that contains at least two elements, the first of which is `maxReadableIndex` and the second of which is therefore greater than `maxReadableIndex` and is thus not `READABLE`.

I've tested this by adjusting the test of `ChainedPlanCache` to more accurately mock a real `RecordStore`'s state. I've also added tests that make sure that plans get selected for if and only if their indexes are in the readable set, which validates the new endpoints on the cache scan.