(Expose Directory Layer as Relational Schema) - Alternative BackingStore (REL-291)
* Put interface in front of FDBRecordStore that can be later replaced with a directory layer based implementation

* Add BackingStore implementation built on top of a `LocatableResolver`

This includes creating a pseudo-meta-data for it that models a directory layer's meta-data. This currently isn't every created by a `Database` object, but tests have been added with a custom database implementation that always creates instances of the new backing store to show that it (essentially) works.

This uses the `LocatableResolver` methods that are currently exposed, and it chooses to make `executeGet` resolve the key. That means that if the database is given a key that isn't already in the table, it will be allocated right then. We could, if we wished, instead make this more like a normal table, and then add methods to allocate missing values on insert (for example). Those methods exist as package-private or protected methods on a `LocatableResolver` at the moment, and we'd have to think a bit more carefully about things like the resolver state. I've left it simpler for now to avoid having to reimplement those features on top of a `LocatableResolver`, but that might be a mistake.