Resolves #2052: Unnecessary recount of records on new stores
This adjusts the logic for handling the record count key so that on new stores, it doesn't have to recount the records. This isn't necessary because with a new store, the store is empty and thus the count key space can be defaulted to having no entries (that is, implicit zeroes for each entry). This also removes an extraneous clear range when the record count key doesn't need to be persisted.

This resolves #2052.