Resolves #2108: Cascades should match record type key field with record type restrictions
In order to execute queries where record type key columns can be used to match type predicates, this modifies the logic around how matches are constructed so that the record type key column can be matched and bound when needed to fix a given type. This then allows for that record type column to be used to narrow down the type, at least for the primary scan.

This resolves #2108.