Resolves #2112: Add a length-limited query plan printer
This adds a new `RecordQueryPlanVisitor` called `PlanStringRepresentation`. It encapsulates the job of constructing the query plan string for a given query plan tree, and it has a parameter to allow for the maximum size of such a string to be capped. This allows for arbitrarily large and complex plans to be packed into a string of fixed size. This can then in turn be used to limit the amount data logged for those larger plans.

An effort was made to keep plan strings the same. Note that the `toString` implementations on the plan objects has been replaced with the new visitor to avoid having two implementations of plan-object-to-explain-string implementations. To test that strings were kept the same, the randomized tester that was added here was run with the old `toString` implementations. Then the implementations were switched over to use the new visitor without changing the test. The exceptions here are:

* Previously, explode plans had an unmatched `]` that appears to have been a typo. In particular, it would generate strings like `explode(@2])`. I've added in a new `[` to match it, so the strings now look like `explode([@2])`
* For `RecordQueryInValuesJoinPlan`, the explain string didn't include whether the values were sorted (ascending or descending). All of the other `InSource`s did, so the new string representation aligns that `InSource` with the other `RecordQueryInJoinPlan` string representations

This resolves #2112.