Resolves #2123: Add unit tests around recreating indexes with new subspace keys
This adds a new unit test to make sure that if an index is recreated with a new subspace key and the old key is left with a `FormerIndex`, then the old index will be cleared out at `checkVersion` time, and operations will use the new subspace key. This ensures that when an index is dropped and recreated in this way, we don't try and use old data for the new index.

This resolves #2123.