Fixes #2136: Lucene query plan toString results in StackOverflow
This adds an explicit implementation of `toString` to the Lucene query plan that calls the right method of a the plan string representation class in order to simulate the old behavior (which was to inherit its `toString` from `RecordQueryIndexPlan`). The other thing that occurred to me was to update `visitDefault` so that it would do something else if we were in `visitDefault` but had a `RecordQueryIndexPlan`. This could either be specific to a `RecordQueryIndexPlan`, or it could walk up the class hierarchy until it found a plan with a known `visit` implementation. I went with the approach I did because it seemed like this approach gives classes outside of `fdb-record-layer-core` the best ability to chose their own `toString` implementations, but there is some waste. In particular, each `LuceneIndexQueryPlan` in the plan tree will create its own `PlanStringRepresentation` class (and underlying `StringBuilder`), but note that these plans don't have any children, so the relative amount that this should cause waste should be relatively contained.

This fixes #2136.