(Use SQLLINE instead of fdb-relational-cli) (REL-464)
*  (Use SQLLINE instead of fdb-relational-cli)

Adds sqlline as the relational command-line. The old CLI and
sqlline coexist currently. The latter will be removed once
folks are good w/ the new sqlline tooling in a follow-on PR.

This PR also adds an embedded JDBC Driver. We now have two JDBC Drivers:
one for embedded access to Relational (added by this PR); and the original
JDBC client for connecting to Relational Servers (whether remote, local or
in-process).

Sqlline works with both.

Both drivers are service loaded so usually you just have to use the
jdbc url to make use of them (no need of specifying driver class).

When sqlline runs, after some work the startup is clean of 'spew' except
for:

 “WARNING: sun.reflect.Reflection.getCallerClass is not supported. This will impact performance.“ Harmless.

This is a harmless log4j2 issue. See https://issues.apache.org/jira/browse/LOG4J2-2537

Spew-removal mostly comes of fixing our logging configuration. RecordLayer
used dictate our log configuration because the recordlayer dependency bundled a
log4j configuration file named 'log4j2-test.properties' on the CLASSPATH which
is the 'pole' position when it comes to configuring logging as far as log4j is
concerned. In this PR we added specifying explicitly our logging configuration
file (to override recordlayer and the other log4j configurations found scattered
along the CLASSPATH in various jars). We do this by passing the system properity
log4j2.configurationFile on JVM startup where it counts -- staring the
Relational Server and running sqlline. Both 'applications' now come w/
accompanying 'conf' directories to host the log4j2 configuration (with more
configurations to come). This makes it so you can mess w/ logging configuration:
edit your configs and restart of your server or sqlline. This PR also
moves to actively exclude transitive includes of slf4j, a logging 'impostor'
as far as Relational is concerned.

We've added some customizations undoing the 'default' username/password challenges
for now until we have implementation and we add some sqlline 'commands':

 * Support for the cascades planner debugger. Run the sqlline command
 named 'plannerdebugger' (or alias 'pd') at the sqlline prompt to
 debug a sql statement:

   !plannerdebugger <SQL_TO_DEBUG>
   e.g. !plannerdebugger select * from databases;
   or shortcut !pd select * from databases;

 * A '!setschema` commmand:

    !setschema CATALOG

Modules have been moved around so fdb-relational-cli is last in line dependent
on all before so embedded and client JDBC Drivers are available to it
(and the yaml-tests are beyond the v-cli). This PR removes the
relational-universal-jdbc module. It had one test class
in it only and its reason for existence goes away until someone asks for a
fat driver jar that can do all of our jdbc dispositions in one jar
(unlikely).

Documentation has been given an edit pointing users at the new cli over
the old.

Still TODO in follow-on:
* Custom outputters for JSON and TXT relational STRUCTS and ARRAYs;
  there is means in sqlline for customizing outputters and at least
  one place-holder is included in this PR but completion is a TODO
  (not needed just yet, not until the following item lands).
* Complete move of yaml-tests from dbstate/old-cli to embedded JDBC Driver.
  All tests but those dependent on the Direct Access API Inserts have
  been moved; completion is gated on a PR in review that is undoing
  DA API Inserts in yaml-tests)
* Removal of old CLI and associated supporting classes when folks are
  good w/ it.

* gradle/libraries.gradle
 Make the original fdb-relational-cli and the incoming sqlline use same jline
 version.

* fdb-relational-api/src/main/java/com/apple/foundationdb/relational/api/SqlTypeNamesSupport.java
 Minor bug fix

* fdb-relational-api/src/main/java/com/apple/foundationdb/relational/api/StructResultSetMetaData.java
 Add an implementation of getColumnTypeName and then a workaround for
 now for the fact that we do not YET do column labels.

* fdb-relational-api/src/main/java/com/apple/foundationdb/relational/api/RelationalDatabaseMetaData.java
 We can now add some default implementations rather than just throw
 exceptions (and stop any further processing).

* fdb-relational-api/src/main/java/com/apple/foundationdb/relational/api/RelationalPreparedStatement.java
* fdb-relational-api/src/main/java/com/apple/foundationdb/relational/api/RelationalStatement.java
 Return no warnings rather than an exception until we implement them
 (returning an exception stops any further processing).

* fdb-relational-api/src/main/java/com/apple/foundationdb/relational/api/RelationalResultSet.java
 Add some default implementaitons in place of throwing an exception
 (returning an exception stops any further processing).

* fdb-relational-api/src/main/java/com/apple/foundationdb/relational/api/RelationalResultSetMetaData.java
 Change how getTableNames works. Act like Oracle (apparently) and return
 an empty String rather than throw an exception and stop any further
 processsing.

* fdb-relational-cli/README.md
 Updated to talk up the new cli.

* fdb-relational-cli/bin/sqlline
 A script to launch sqlline in a build dev environment; it finds the
 script generated by the gradle application plugin and lanuches it.

* fdb-relational-cli/build.gradle
 Bundle sqlline and add generation of start script for sqlline that
 factors the Relational context.

* fdb-relational-cli/src/main/dist/conf/log4j2.xml
 log4j2 configuration for sqline

* fdb-relational-cli/src/main/java/com/apple/foundationdb/relational/cli/sqlline/Customize.java
 Relational customization of sqlline handling.
 Adds alternate json and text formatters and bundles in our
 plannerdebugger handler.

* fdb-relational-cli/src/main/java/sqlline/PlannerDebuggerCommandHandler.java
 sqlline command to launch the cascades debugger.

* fdb-relational-cli/src/test/java/com/apple/foundationdb/relational/cli/SQLLineTest.java
 Test that sqlline is basically working (useful for debugging how
 sqlline is working internally).

* fdb-relational-cli/src/test/java/com/apple/foundationdb/relational/cli/RelationalReplTest.java
 Disabled. Its going away.

* fdb-relational-core/build.gradle
 Tell serviceloader about the embedded JDBC Driver.
     private final EmbeddedRelationalEngine engine;

* fdb-relational-core/src/main/java/com/apple/foundationdb/relational/jdbc/JDBCEmbedDriver.java
 Embedded Driver.

* fdb-relational-core/src/main/java/com/apple/foundationdb/relational/recordlayer/RecordLayerResultSet.java
 isCLosed implementation.

* fdb-relational-server/bin/relational
 Add a script to start server in built dev env. Same as one in
 v-cli/bin.

* fdb-relational-server/src/main/dist/conf/log4j2.xml
 log4j2 config for server (as opposed to what we had before where config
 was built into the jar).

* yaml-tests/src/test/java/EmbeddedJDBCYamlIntegrationTests.java
 Run yaml test suite against the embedded JDBC Driver (all tests
 but the ones that do Direct Access API Insert).

* Move log4j2 version back to 2.18... Save update to 2.20 till a separate issue

* Address Youssef review

Signed-off-by: yhatem
Signed-off-by: bfines