(Do not reconstruct TypeRepository with every query execution) (REL-479)
This puts the TypeRepository in the PhysicalQueryPlan. That allows the same type repository to be used for multiple executions of the same query plan, which hopefully saves time during query execution. For uncached plans, this should be slightly more efficient, as the old path required constructing the type respository during planning and then reconstructing it during execution, whereas the new path constructs one type repository for planning and then updates it with any new types. For cached plans, the code now allows the plan to skip construction of the `TypeRepository` on cached plans entirely.