(Memoize QueryCacheKey hash code) (REL-478)
This is one potential performance issue I saw when I was digging into plan cache performance. I'm not sure how much of a difference this will really make, but given that the query cache key is evaluated at least once for every cache lookup and potentially multiple times for each cache entry, it seemed that eagerly evaluating this and memoizing it would make sense.