Add additional tests for empty table aggregation (REL-487)
This PR adds additional tests for COUNT(*), COUNT(col) and SUM(col) on empty tables.
This also checks that the behavior is the same for tables that have always been empty and for tables that used to contain elements but were cleared. 4 follow-up issues were filed to track bug fixing and future work.
Also part of this PR:
Simplification of the explain yaml command, it no longer require a resultSet as a result but a simple string instead
Introducing a new YAML command: explainContains to check for substrings in the plan